fix!: ensure we only pull artifacts when running full or fast layer validation (#256)
## Issue
Resolves
https://github.com/validator-labs/validator-plugin-oci/issues/242

## Description
This PR accomplishes a few things:
1. Ensures we only pull images when running full or fast layer
validation
2. Rewords details to reflect the above

It does this by introducing a `ValidationType` which can be set to
`full`, `fast`, or `none`.
At the host level, `ValidationType` is defaulted to `none`. When set on
the artifact level, you can override the value configured on the host.
This also means that `SkipLayerValidation` is no longer a supported
field.